api/src/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
gunicorn==21.2.0
requests==2.31.0
types-requests==2.31.0.2
pytest==7.4.0
httpx==0.24.1
fastapi-cache2==0.2.1
redis==4.6.0
asyncpg==0.28.0
fastapi-users==12.1.1
fastapi-users-db-sqlalchemy==6.0.1
pydantic==2.3.0
pydantic-settings==2.0.3
SQLAlchemy==2.0.20
databases==0.6.2
